Swap locations of two functions
`set_particle_element_ids` is moved to element.py since the element id
mapping is based on what type of element the particles lie in. On the
other hand, `update_natural_coords` was moved to particle.py since it is
an intrinsic property of the particles and is just updated based on the
elements passed as argument.