Implement correct 2d JIT solve!
To correct the 2D JIT solve with correct volume computation, 2 major
changes were made:
1. The elements class constructor now takes the total number of elements
as an argument along with number of elements in both directions. This is
so that `jnp.bincount` can work correctly without having to pass a
tracer value to its `length` parameter.
2. Update the `MPMExplicit` flattening/unflattening to include the
`velocity_update` parameter to make it work with JIT.